% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_lib.R
\name{accumulate_interaction_effects}
\alias{accumulate_interaction_effects}
\title{Accumulate responses of group membership interactions}
\usage{
accumulate_interaction_effects(group_effects)
}
\arguments{
\item{group_effects}{A list of group effects, e.g. as returned by \code{draw_group_effects}}
}
\value{
A dataframe with regressors, groups, and ey values suitable for
joining with raw data.
}
\description{
Accumulate responses of group membership interactions
}
