% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_lib.R
\name{aggregate_simuilation_data}
\alias{aggregate_simuilation_data}
\title{Aggregate simulation data by group.}
\usage{
aggregate_simuilation_data(sim_data)
}
\arguments{
\item{sim_data}{Output of \code{simulate_survey_data}}
}
\value{
A list containing aggregated and joined simulation data
}
\description{
Aggregate simulation data by group.
}
