% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_lib.R
\name{convert_to_consistent_factors}
\alias{convert_to_consistent_factors}
\title{Encode two vectors to factors ensuring that the levels are the same.}
\usage{
convert_to_consistent_factors(vec1, vec2)
}
\arguments{
\item{vec1}{A vector to be converted to a factor}

\item{vec2}{A vector to be converted to a factor}
}
\value{
A list containing consistent factor versions of vec1 and vec2
}
\description{
Encode two vectors to factors ensuring that the levels are the same.
}
