% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_lib.R
\name{draw_conditional_binary_data}
\alias{draw_conditional_binary_data}
\title{Draw binary data with e_y_new, conditional on original binary draws
y_orig and an estimate of their expectation, e_y.}
\usage{
draw_conditional_binary_data(y_orig, e_y, e_y_new, unif_base = NULL)
}
\arguments{
\item{y_orig}{A vector of binary {0, 1} draws}

\item{e_y}{A vector of estimated expectatoins of y_orig}

\item{e_y_new}{A vector of expectations of the new binary data}

\item{unif_base}{(Optional) Uniform \link{0,1} random numbers to condition on.
If unspecified, new draws are taken.}
}
\value{
A list containing
\itemize{
\item ytil: A new draw of binary random variables.  Marginally, ytil has
mean e_y_new if y_orig had mean e_y
\item unif_base: The uniform random numbers used for the analysis
\item unif_cond: A rescaled version of unif_base to represent a
draw of u | y.
}
}
\description{
Draw binary data with e_y_new, conditional on original binary draws
y_orig and an estimate of their expectation, e_y.
}
