% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_lib.R
\name{draw_correlated_groups}
\alias{draw_correlated_groups}
\title{Draw correlated group memberships with a censored normal}
\usage{
draw_correlated_groups(n_obs, latent_mean, latent_cov)
}
\arguments{
\item{n_obs}{The number of rows to draw}

\item{latent_mean}{A length \code{p} mean of the latent variable}

\item{latent_cov}{A \code{p} by \code{p} covariance of the latent variable}
}
\value{
An \code{nobs} by \code{p} matrix of group membership.  A row
is a member of the group if the corresponding component of its
latent normal is greater than zero.
}
\description{
Draw correlated group memberships with a censored normal
}
