% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_lib.R
\name{draw_group_effects}
\alias{draw_group_effects}
\title{Draw group membership effect sizes}
\usage{
draw_group_effects(n_groups, degree)
}
\arguments{
\item{n_groups}{The number of distinct groups}

\item{degree}{The degree of interaction}
}
\value{
A list containing a number of dataframes.
\itemize{
\item beta_df: The coefficients used to generate the effects
\item regressor_df: The mapping between regressors (group interactions) and categories
\item group_df: The mapping between individual group membership and categories
\item reg_form: The formula string used to generate the model matrix.
\item g_cols: The names of the groups.
The dataframe shared columns are as follows:
\item s: An integer identifying each unique combination of groups
\item name: A readable name of a regressor, i.e. a group interaction term
}
}
\description{
Draw group membership effect sizes
}
