% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_lib.R
\name{draw_response}
\alias{draw_response}
\title{Draw binary responses for a particular set of group memberships}
\usage{
draw_response(g_matrix, effect_df)
}
\arguments{
\item{g_matrix}{A matrix of group memberships, e.g. as returned by \code{draw_correlated_groups}}

\item{effect_df}{A dataframe containing the columns in \code{g_matrix} as well as
expected response \code{ey}, e.g. as returned by \code{accumulate_interaction_effects}.}
}
\value{
A dataframe containing group memberships \code{s} and responses \code{y} corresponding
to the rows of \code{g_matrix}.
}
\description{
Draw binary responses for a particular set of group memberships
}
