% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_lib.R
\name{get_consistent_regressors}
\alias{get_consistent_regressors}
\title{Create regressor matrices with consistent encodings.}
\usage{
get_consistent_regressors(form, df1, df2, frame = FALSE)
}
\arguments{
\item{form}{A model formula}

\item{df1}{A dataframe for model matrix}

\item{df2}{A dataframe for model matrix}

\item{frame}{Use model.frame rather than model.matrix.}
}
\value{
A list containing regressor dataframes x1 and x2 for df1 and df2
respectively with consistent encodings.  NAs will be included.
}
\description{
Create regressor matrices with consistent encodings.
}
