% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrplew_mcmc_lib.R
\name{get_mrp_draws}
\alias{get_mrp_draws}
\title{Get mrp draws MCMC estimators.}
\usage{
get_mrp_draws(yhat_pop_draws, pop_w = NULL)
}
\arguments{
\item{yhat_pop_draws}{MCMC draws x Population size matrix of yhat posterior draws}

\item{pop_w}{Optional.  The weight given to each row of pop_df.  Defaults to ones.}
}
\value{
Draws from the MrP estimate.
}
\description{
Get mrp draws MCMC estimators.
}
