% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_lib.R
\name{get_mrplew_logistic_brms}
\alias{get_mrplew_logistic_brms}
\title{Get mrplew weights for the brms logistic MCMC estimator.}
\usage{
get_mrplew_logistic_brms(
  brms_post,
  survey_df,
  pop_df,
  pop_w = NULL,
  save_draws = FALSE
)
}
\arguments{
\item{brms_post}{The output of \code{brm(..., survey_df, family=binomial(link="logit"))}}

\item{survey_df}{The survey dataframe}

\item{pop_df}{The population dataframe}

\item{pop_w}{Optional.  The weight given to each row of pop_df.  Defaults to ones.}

\item{save_draws}{Optional.  If true, save the posterior predictions for re-use.}
}
\value{
Draws from the MrP estimate, and the weight vector
whose n-th entry is d E\link{MrP | X, Y} / d y_n.
}
\description{
Get mrplew weights for the brms logistic MCMC estimator.
}
