% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrplew_mcmc_lib.R
\name{get_mrplew_mcmc}
\alias{get_mrplew_mcmc}
\title{Get mrplew weights for MCMC estimators.}
\usage{
get_mrplew_mcmc(mrp_draws, dloglikdy_survey_draws)
}
\arguments{
\item{mrp_draws}{MCMC draws of the MrP estimator}

\item{dloglikdy_survey_draws}{MCMC draws x Survey size matrix of draws of d log p(y_i | theta) / dy_i}
}
\value{
Draws from the MrP estimate, and the weight vector
whose i-th entry is d E\link{MrP | X, Y} / d y_i where i indexes survey observations.
}
\description{
Get mrplew weights for MCMC estimators.
}
