% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_lib.R
\name{get_ols_mcmc_weights}
\alias{get_ols_mcmc_weights}
\title{Get mrplew weights for the brms logistic MCMC estimator.}
\usage{
get_ols_mcmc_weights(
  brms_post,
  survey_df,
  pop_df,
  pop_frac = NULL,
  re_formula = NULL,
  allow_new_levels = FALSE
)
}
\arguments{
\item{brms_post}{The output of \code{brm(..., survey_df, family=gaussian())}}

\item{survey_df}{The survey dataframe}

\item{pop_df}{The population dataframe}

\item{pop_frac}{Optional.  The weight given to each row of pop_df.  Defaults to ones.}
}
\value{
Draws from the MrP estimate, and the weight vector
whose n-th entry is d E\link{MrP | X, Y} / d y_n.
}
\description{
Get mrplew weights for the brms logistic MCMC estimator.
}
