% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lib.R
\name{get_response}
\alias{get_response}
\title{Get the response variable (y) from the posterior.
I don't see this use clearly documented, so I want to factor it out
for testing.}
\usage{
get_response(post)
}
\arguments{
\item{post}{A brms posterior}
}
\value{
The numeric response variable used for the posterior fitting
}
\description{
Get the response variable (y) from the posterior.
I don't see this use clearly documented, so I want to factor it out
for testing.
}
