% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_lib.R
\name{safe_model_matrix}
\alias{safe_model_matrix}
\title{A version of model.matrix that respects na.action.}
\usage{
safe_model_matrix(form, df, na.action = na.pass)
}
\arguments{
\item{form}{A formula}

\item{df}{A dataframe}
}
\value{
The output of model.matrix(form, df) but respecting na.action.
}
\description{
A version of model.matrix that respects na.action.
}
