% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_lib.R
\name{simulate_survey_data}
\alias{simulate_survey_data}
\title{Draw survey data with some reasonable defaults}
\usage{
simulate_survey_data(n_groups, n_obs, n_obs_pop, degree)
}
\arguments{
\item{n_groups}{The number of distinct groups}

\item{n_obs}{The number of population observations}

\item{degree}{The maximum degree of true interactions}
}
\value{
A list of simualted data
}
\description{
Draw survey data with some reasonable defaults
}
