% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_opt_lib.R
\name{get_mrplew_lm}
\alias{get_mrplew_lm}
\title{Get mrplew weights for the OLS estimator.}
\usage{
get_mrplew_lm(lm_fit, survey_df, pop_df, pop_w = NULL)
}
\arguments{
\item{lm_fit}{The output of \code{lm}}

\item{survey_df}{The survey dataframe}

\item{pop_df}{The population dataframe}
}
\value{
The MrP OLS estimate for sanity checking, and the weight vector
whose n-th entry is d MrP / d y_n.
}
\description{
Get mrplew weights for the OLS estimator.
}
