% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_opt_lib.R
\name{get_mrplew_logistic_glm}
\alias{get_mrplew_logistic_glm}
\title{Get mrplew weights for the logistic estimator.}
\usage{
get_mrplew_logistic_glm(logit_fit, survey_df, pop_df, pop_w = NULL)
}
\arguments{
\item{logit_fit}{The output of \code{glm(..., survey_df, family=binomial(link="logit"))}}

\item{survey_df}{The survey dataframe}

\item{pop_df}{The population dataframe}
}
\value{
The MrP Logistic estimate for sanity checking, and the weight vector
whose n-th entry is d MrP / d y_n.
}
\description{
Get mrplew weights for the logistic estimator.
}
