% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrplew_mcmc_lib.R
\name{get_mrplew_mcmc}
\alias{get_mrplew_mcmc}
\title{Get mrplew weights for MCMC estimators.}
\usage{
get_mrplew_mcmc(yhat_pop_draws, dloglikdy_survey_draws, pop_w = NULL)
}
\arguments{
\item{yhat_pop_draws}{MCMC draws x Population size matrix of yhat posterior draws}

\item{dloglikdy_survey_draws}{MCMC draws x Survey size matrix of draws of d log p(y_i | theta) / dy_i}

\item{pop_w}{Optional.  The weight given to each row of pop_df.  Defaults to ones.}
}
\value{
Draws from the MrP estimate, and the weight vector
whose i-th entry is d E\link{MrP | X, Y} / d y_i where i indexes survey observations.
}
\description{
Get mrplew weights for MCMC estimators.
}
