% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_mcmc_lib.R
\name{get_ols_mcmc_weights}
\alias{get_ols_mcmc_weights}
\title{Get mrplew weights for the logistic MCMC estimator.}
\usage{
get_ols_mcmc_weights(
  lin_post,
  survey_df,
  pop_df,
  pop_w = NULL,
  re_formula = NULL,
  allow_new_levels = FALSE
)
}
\arguments{
\item{lin_post}{The output of \code{brm(..., survey_df, family=gaussian())}}

\item{survey_df}{The survey dataframe}

\item{pop_df}{The population dataframe}

\item{pop_w}{Optional.  The weight given to each row of pop_df.  Defaults to ones.}

\item{re_formula}{Optional.  Formula containing group-level effects to be considered in the prediction. If \code{NULL} (default), include all group-level effects; if \code{NA}, include no group-level effects.}

\item{allow_new_levels}{Optional.  If true, allow new levels of group-level effects in prediction stage.}
}
\value{
Draws from the MrP estimate, and the weight vector
whose n-th entry is d E\link{MrP | X, Y} / d y_n.
}
\description{
Get mrplew weights for the logistic MCMC estimator.
}
